Move servo setup to a config file.
This patch moves the controller
addresses to a config file, along
with the servo PWM range of motion.

This allows one to use any number of
servo controllers on the i2c bus,
and also makes it easier to
customize the servo calibration.

This patch also includes a Servo
Calibration GUI, which makes it
much easier to calibrate both
the controller and channel, as well
as the min and max pwm for each servo.